created a separate file index.ios.tsx
